requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
lorem-text==3.0
lxml==6.0.0
packaging==25.0
pillow==11.3.0
playwright
pluggy==1.6.0
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.2
pyperclip==1.9.0
PyScreeze==1.0.1
pytest==8.4.1
pytest-base-url==2.1.0
pytest-playwright==0.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytweening==1.2.0
pytz==2025.2
requests==2.32.4
